+++ src/hew_back/tbls/__product_table.py
@@ -3,22 +3,17 @@ import uuid
 from typing import Union, List
 
 import sqlalchemy
-from sqlalchemy import Integer
 
-from sqlalchemy import Column, String, DateTime, UUID
+from sqlalchemy import Column, String, DateTime, UUID, select, update
 from sqlalchemy import select, or_, func
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from hew_back import tbls
 from hew_back.db import BaseTable
 from hew_back.util import OrderDirection
 
+from hew_back import tbls
 
-# from asyncpg.pgproto.pgproto import UUID
-
-
-# from hew_back import table
-
+from zoneinfo import ZoneInfo
 
 class ProductTable(BaseTable):
     __tablename__ = 'TBL_PRODUCT'
@@ -175,5 +170,43 @@ class ProductTable(BaseTable):
         products = result.scalars().all()
 
         return products
+
+
+    @staticmethod
+    async def get_cart_products(
+            session: AsyncSession,
+            user_id: tbls.UserTable,
+    ) -> List['ProductTable']:
+        # 該当するユーザー
+        stmt = (
+            select(tbls.ProductTable)
+            .join(tbls.CartProductTable, tbls.ProductTable.product_id == tbls.CartProductTable.product_id)
+            .join(tbls.CartTable, tbls.CartProductTable.cart_id == tbls.CartTable.cart_id)
+            .where(tbls.CartTable.user_id == user_id)
+            .where(tbls.CartTable.purchase_date == None)
+        )
+        result = await session.execute(stmt)
+        product_cart = result.scalars().all()
+        return list(product_cart)
+
+    @staticmethod
+    async def cart_buy(
+        session: AsyncSession,
+        user_id: tbls.UserTable,
+    ):
+        subquery = (
+            select(tbls.CartTable.cart_id)
+            .where(tbls.CartTable.user_id == user_id)
+            .where(tbls.CartTable.purchase_date == None)
+        )
+
+        stmt = (
+            update(tbls.CartTable)
+            .where(tbls.CartTable.cart_id.in_(subquery))
+            .values(purchase_date=datetime.datetime.now(ZoneInfo("Asia/Tokyo")).replace(tzinfo=None))
+        )
+
+        await session.execute(stmt)
+        await session.commit()
