+++ src/test/base.py
@@ -1,6 +1,7 @@
 import pydantic
 from fastapi import FastAPI
 from httpx import AsyncClient
+from pygments.lexer import default
 
 
 class Client:
@@ -21,11 +22,14 @@ class Client:
         result = await self.client.post(path, content=json_data.model_dump_json(), headers=headers)
         return result
 
-    async def put(self, path: str, json_data: pydantic.BaseModel, token: str | None = None):
+    async def put(self, path: str, json_data: pydantic.BaseModel | None = None, token: str | None = None):
         headers = {}
         if token is not None:
             headers["Authorization"] = f"Bearer {token}"
-        result = await self.client.put(path, content=json_data.model_dump_json(), headers=headers)
+        if json_data is None:
+            result = await self.client.put(path, headers=headers)
+        else:
+            result = await self.client.put(path, content=json_data.model_dump_json(), headers=headers)
         return result
 
     async def delete(self, path: str, token: str | None = None):
